be/requirements.txt
asgiref==3.5.2
cssselect2==0.6.0
Django==4.0.5
#django-filer==2.2.1
django-js-asset==2.0.0
django-log-request-id==2.0.0
django-mptt==0.13.4
django-nested-admin==3.4.0
#django-nested-inline==0.4.5
#django-nonrelated-inlines==0.2
#django-polymorphic==3.1.0
djangorestframework==3.13.1
lxml==4.9.1
Pillow==9.2.0
pytz==2022.1
reportlab==3.6.10
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tzdata==2022.1
Unidecode==1.1.2
webencodings==0.5.1
PyYAML==6.0